# api/docs/swagger.yaml
definitions:
  model.Customer:
    properties:
      address_line1:
        type: string
      address_line2:
        type: string
      city:
        type: string
      contact_first_name:
        type: string
      contact_last_name:
        type: string
      country:
        type: string
      credit_limit:
        type: integer
      customer_name:
        type: string
      customer_number:
        type: integer
      email:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      sales_rep_employee_number:
        type: integer
      state:
        type: string
    type: object
  model.Employee:
    properties:
      email:
        type: string
      employee_number:
        type: integer
      extension:
        type: string
      first_name:
        type: string
      job_title:
        type: string
      last_name:
        type: string
      office_code:
        type: string
      reports_to:
        type: integer
    type: object
  model.ErrValidation:
    properties:
      invalid_params:
        items:
          $ref: '#/definitions/model.InvalidParam'
        type: array
    type: object
  model.InvalidParam:
    properties:
      name:
        type: string
      reason:
        type: string
    type: object
  model.Office:
    properties:
      address_line:
        type: string
      address_line2:
        type: string
      city:
        type: string
      country:
        type: string
      office_code:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      state:
        type: string
      territory:
        type: string
    type: object
...
info:
  contact:
    email: Martsoftfilmz@gmail.com
    name: Martsoft Inc
  description: This is the API E-Commerce businesses.
  title: Martsoft Inc E-Commerce API 2.0
  version: "1.0"
paths:
  /customers/customers:
    get:
      description: Gets all customers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve all customers
      tags:
      - Customer
  /customers/delete:
    delete:
      description: Deletes a single customer
      parameters:
      - description: email mandatory
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Deletes a single customer
      tags:
      - Customer
  /customers/get:
    get:
      description: Fetch a single customer
      parameters:
      - description: email mandatory
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve a single customer
      tags:
      - Customer
  /customers/update:
    patch:
      description: updates a single customer
      parameters:
      - description: customer_number
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Updates a  customer
      tags:
      - Customer
  /employees/delete:
    delete:
      description: deleted a single Employee
      parameters:
      - description: employee_number mandatory
        in: query
        name: employee_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: deletes a single Employee
      tags:
      - Employees
  /employees/employees:
    get:
      description: Gets all Employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve all Employees
      tags:
      - Employees
  /employees/get:
    get:
      description: Fetch a single Employee
      parameters:
      - description: employee_number mandatory
        in: query
        name: employee_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve a single Employee
      tags:
      - Employees
  /employees/update:
    patch:
      description: updates a single Employee
      parameters:
      - description: employee_number
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/model.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Updates a  Employee
      tags:
      - Employees
  /offices/delete_office:
    delete:
      description: Deletes a single Office
      parameters:
      - description: office_code mandatory
        in: query
        name: office_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Office'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Deletes a single Office
      tags:
      - Offices
  /offices/get:
    get:
      description: Fetch a single Office
      parameters:
      - description: office_code mandatory
        in: query
        name: office_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Office'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve a single Office
      tags:
      - Offices
  /offices/offices:
    get:
      description: Gets all Offices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Office'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Retrieve all Offices
      tags:
      - Offices
  /offices/update:
    patch:
      description: updates a single Office
      parameters:
      - description: office_code
        in: body
        name: Office
        required: true
        schema:
          $ref: '#/definitions/model.Office'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Office'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrValidation'
      summary: Updates a  Office
      tags:
      - Offices
...
schemes:
- http
swagger: "2.0"